# docs/swagger.yaml
definitions:
...
  cloudevent.CloudEvent-encoding_json_RawMessage:
    properties:
      data:
        description: Data contains domain-specific information about the event.
        items:
          type: integer
        type: array
      datacontenttype:
        description: |-
          DataContentType is an optional MIME type for the data field. We almost
          always serialize to JSON and in that case this field is implicitly
          "application/json".
        type: string
      dataschema:
        description: DataSchema is an optional URI pointing to a schema for the data
          field.
        type: string
      dataversion:
        description: DataVersion is the version of the data type.
        type: string
      id:
        description: |-
          ID is an identifier for the event. The combination of ID and Source must
          be unique.
        type: string
      producer:
        description: Producer is a specific instance, process or device that creates
          the data structure describing the CloudEvent.
        type: string
      source:
        description: Source is the context in which the event happened. In a distributed
          system it might consist of multiple Producers.
        type: string
      specversion:
        description: |-
          SpecVersion is the version of CloudEvents specification used.
          This is always hardcoded "1.0".
        type: string
      subject:
        description: |-
          Subject is an optional field identifying the subject of the event within
          the context of the event producer. In practice, we always set this.
        type: string
      time:
        description: |-
          Time is an optional field giving the time at which the event occurred. In
          practice, we always set this.
        type: string
      type:
        description: |-
          Type describes the type of event. It should generally be a reverse-DNS
          name.
        type: string
    type: object
  github_com_DIMO-Network_odometer-attester_internal_client_identity.GraphQLResponse:
    properties:
      data:
        $ref: '#/definitions/github_com_DIMO-Network_odometer-attester_internal_client_identity.dataField'
      errors:
        items:
          $ref: '#/definitions/github_com_DIMO-Network_odometer-attester_internal_client_identity.graphQLError'
        type: array
    type: object
  github_com_DIMO-Network_odometer-attester_internal_client_identity.dataField:
    properties:
      vehicle:
        $ref: '#/definitions/github_com_DIMO-Network_odometer-attester_internal_client_identity.vehicleField'
    type: object
...
  github_com_DIMO-Network_odometer-attester_internal_client_identity.graphQLError:
    properties:
      message:
        type: string
    type: object
...
  internal_app.codeResp:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  nitrite.Document:
    properties:
      cabundle:
        items:
          items:
            type: integer
          type: array
        type: array
      certificate:
        items:
          type: integer
        type: array
      digest:
        type: string
      module_id:
        type: string
      nonce:
        items:
          type: integer
        type: array
      pcrs:
        additionalProperties:
          items:
            type: integer
          type: array
        type: object
      public_key:
        items:
          type: integer
        type: array
      timestamp:
        type: integer
      user_data:
        items:
          type: integer
        type: array
    type: object
  nitrite.Result:
    properties:
      certificates:
        description: Certificates contains all of the certificates except the root.
        items:
          $ref: '#/definitions/x509.Certificate'
        type: array
      cose_sign1:
        description: |-
          COSESign1 contains the COSE Signature Structure which was used to
          calculate the `Signature`.
        items:
          type: integer
        type: array
      document:
        allOf:
        - $ref: '#/definitions/nitrite.Document'
        description: Document contains the attestation document.
      payload:
        description: Payload section from the COSE Sign1 payload.
        items:
          type: integer
        type: array
      protected:
        description: Protected section from the COSE Sign1 payload.
        items:
          type: integer
        type: array
      signature:
        description: Signature section from the COSE Sign1 payload.
        items:
          type: integer
        type: array
      signature_ok:
        description: |-
          SignatureOK designates if the signature was OK (but certificate could be
          invalid, not trusted, expired, etc.)
        type: boolean
      unprotected:
        description: Unprotected section from the COSE Sign1 payload.
        items:
          type: integer
        type: array
    type: object
...
  x509.Certificate:
    properties:
      authorityKeyId:
        items:
          type: integer
        type: array
      basicConstraintsValid:
        description: |-
          BasicConstraintsValid indicates whether IsCA, MaxPathLen,
          and MaxPathLenZero are valid.
        type: boolean
      crldistributionPoints:
        description: CRL Distribution Points
        items:
          type: string
        type: array
      dnsnames:
        description: |-
          Subject Alternate Name values. (Note that these values may not be valid
          if invalid values were contained within a parsed certificate. For
          example, an element of DNSNames may not be a valid DNS domain name.)
        items:
          type: string
        type: array
      emailAddresses:
        items:
          type: string
        type: array
      excludedDNSDomains:
        items:
          type: string
        type: array
      excludedEmailAddresses:
        items:
          type: string
        type: array
      excludedIPRanges:
        items:
          $ref: '#/definitions/net.IPNet'
        type: array
      excludedURIDomains:
        items:
          type: string
        type: array
      extKeyUsage:
        description: Sequence of extended key usages.
        items:
          $ref: '#/definitions/x509.ExtKeyUsage'
        type: array
      extensions:
        description: |-
          Extensions contains raw X.509 extensions. When parsing certificates,
          this can be used to extract non-critical extensions that are not
          parsed by this package. When marshaling certificates, the Extensions
          field is ignored, see ExtraExtensions.
        items:
          $ref: '#/definitions/pkix.Extension'
        type: array
      extraExtensions:
        description: |-
          ExtraExtensions contains extensions to be copied, raw, into any
          marshaled certificates. Values override any extensions that would
          otherwise be produced based on the other fields. The ExtraExtensions
          field is not populated when parsing certificates, see Extensions.
        items:
          $ref: '#/definitions/pkix.Extension'
        type: array
      inhibitAnyPolicy:
        description: |-
          InhibitAnyPolicy and InhibitAnyPolicyZero indicate the presence and value
          of the inhibitAnyPolicy extension.

          The value of InhibitAnyPolicy indicates the number of additional
          certificates in the path after this certificate that may use the
          anyPolicy policy OID to indicate a match with any other policy.

          When parsing a certificate, a positive non-zero InhibitAnyPolicy means
          that the field was specified, -1 means it was unset, and
          InhibitAnyPolicyZero being true mean that the field was explicitly set to
          zero. The case of InhibitAnyPolicy==0 with InhibitAnyPolicyZero==false
          should be treated equivalent to -1 (unset).
        type: integer
      inhibitAnyPolicyZero:
        description: |-
          InhibitAnyPolicyZero indicates that InhibitAnyPolicy==0 should be
          interpreted as an actual maximum path length of zero. Otherwise, that
          combination is interpreted as InhibitAnyPolicy not being set.
        type: boolean
      inhibitPolicyMapping:
        description: |-
          InhibitPolicyMapping and InhibitPolicyMappingZero indicate the presence
          and value of the inhibitPolicyMapping field of the policyConstraints
          extension.

          The value of InhibitPolicyMapping indicates the number of additional
          certificates in the path after this certificate that may use policy
          mapping.

          When parsing a certificate, a positive non-zero InhibitPolicyMapping
          means that the field was specified, -1 means it was unset, and
          InhibitPolicyMappingZero being true mean that the field was explicitly
          set to zero. The case of InhibitPolicyMapping==0 with
          InhibitPolicyMappingZero==false should be treated equivalent to -1
          (unset).
        type: integer
      inhibitPolicyMappingZero:
        description: |-
          InhibitPolicyMappingZero indicates that InhibitPolicyMapping==0 should be
          interpreted as an actual maximum path length of zero. Otherwise, that
          combination is interpreted as InhibitAnyPolicy not being set.
        type: boolean
      ipaddresses:
        items:
          items:
            type: integer
          type: array
        type: array
      isCA:
        type: boolean
      issuer:
        $ref: '#/definitions/pkix.Name'
      issuingCertificateURL:
        items:
          type: string
        type: array
      keyUsage:
        $ref: '#/definitions/x509.KeyUsage'
      maxPathLen:
        description: |-
          MaxPathLen and MaxPathLenZero indicate the presence and
          value of the BasicConstraints' "pathLenConstraint".

          When parsing a certificate, a positive non-zero MaxPathLen
          means that the field was specified, -1 means it was unset,
          and MaxPathLenZero being true mean that the field was
          explicitly set to zero. The case of MaxPathLen==0 with MaxPathLenZero==false
          should be treated equivalent to -1 (unset).

          When generating a certificate, an unset pathLenConstraint
          can be requested with either MaxPathLen == -1 or using the
          zero value for both MaxPathLen and MaxPathLenZero.
        type: integer
      maxPathLenZero:
        description: |-
          MaxPathLenZero indicates that BasicConstraintsValid==true
          and MaxPathLen==0 should be interpreted as an actual
          maximum path length of zero. Otherwise, that combination is
          interpreted as MaxPathLen not being set.
        type: boolean
      notAfter:
        description: Validity bounds.
        type: string
      notBefore:
        description: Validity bounds.
        type: string
      ocspserver:
        description: RFC 5280, 4.2.2.1 (Authority Information Access)
        items:
          type: string
        type: array
      permittedDNSDomains:
        items:
          type: string
        type: array
      permittedDNSDomainsCritical:
        description: Name constraints
        type: boolean
      permittedEmailAddresses:
        items:
          type: string
        type: array
      permittedIPRanges:
        items:
          $ref: '#/definitions/net.IPNet'
        type: array
      permittedURIDomains:
        items:
          type: string
        type: array
      policies:
        description: |-
          Policies contains all policy identifiers included in the certificate.
          See CreateCertificate for context about how this field and the PolicyIdentifiers field
          interact.
          In Go 1.22, encoding/gob cannot handle and ignores this field.
        items:
          $ref: '#/definitions/x509.OID'
        type: array
      policyIdentifiers:
        description: |-
          PolicyIdentifiers contains asn1.ObjectIdentifiers, the components
          of which are limited to int32. If a certificate contains a policy which
          cannot be represented by asn1.ObjectIdentifier, it will not be included in
          PolicyIdentifiers, but will be present in Policies, which contains all parsed
          policy OIDs.
          See CreateCertificate for context about how this field and the Policies field
          interact.
        items:
          items:
            type: integer
          type: array
        type: array
      policyMappings:
        description: PolicyMappings contains a list of policy mappings included in
          the certificate.
        items:
          $ref: '#/definitions/x509.PolicyMapping'
        type: array
      publicKey: {}
      publicKeyAlgorithm:
        $ref: '#/definitions/x509.PublicKeyAlgorithm'
      raw:
        description: Complete ASN.1 DER content (certificate, signature algorithm
          and signature).
        items:
          type: integer
        type: array
      rawIssuer:
        description: DER encoded Issuer
        items:
          type: integer
        type: array
      rawSubject:
        description: DER encoded Subject
        items:
          type: integer
        type: array
      rawSubjectPublicKeyInfo:
        description: DER encoded SubjectPublicKeyInfo.
        items:
          type: integer
        type: array
      rawTBSCertificate:
        description: Certificate part of raw ASN.1 DER content.
        items:
          type: integer
        type: array
      requireExplicitPolicy:
        description: |-
          RequireExplicitPolicy and RequireExplicitPolicyZero indicate the presence
          and value of the requireExplicitPolicy field of the policyConstraints
          extension.

          The value of RequireExplicitPolicy indicates the number of additional
          certificates in the path after this certificate before an explicit policy
          is required for the rest of the path. When an explicit policy is required,
          each subsequent certificate in the path must contain a required policy OID,
          or a policy OID which has been declared as equivalent through the policy
          mapping extension.

          When parsing a certificate, a positive non-zero RequireExplicitPolicy
          means that the field was specified, -1 means it was unset, and
          RequireExplicitPolicyZero being true mean that the field was explicitly
          set to zero. The case of RequireExplicitPolicy==0 with
          RequireExplicitPolicyZero==false should be treated equivalent to -1
          (unset).
        type: integer
      requireExplicitPolicyZero:
        description: |-
          RequireExplicitPolicyZero indicates that RequireExplicitPolicy==0 should be
          interpreted as an actual maximum path length of zero. Otherwise, that
          combination is interpreted as InhibitAnyPolicy not being set.
        type: boolean
      serialNumber:
        $ref: '#/definitions/big.Int'
      signature:
        items:
          type: integer
        type: array
      signatureAlgorithm:
        $ref: '#/definitions/x509.SignatureAlgorithm'
      subject:
        $ref: '#/definitions/pkix.Name'
      subjectKeyId:
        items:
          type: integer
        type: array
      unhandledCriticalExtensions:
        description: |-
          UnhandledCriticalExtensions contains a list of extension IDs that
          were not (fully) processed when parsing. Verify will fail if this
          slice is non-empty, unless verification is delegated to an OS
          library which understands all the critical extensions.

          Users can access these extensions using Extensions and can remove
          elements from this slice if they believe that they have been
          handled.
        items:
          items:
            type: integer
          type: array
        type: array
      unknownExtKeyUsage:
        description: Encountered extended key usages unknown to this package.
        items:
          items:
            type: integer
          type: array
        type: array
      uris:
        items:
          $ref: '#/definitions/url.URL'
        type: array
      version:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is the API documentation for the Odometer Attester service
  title: Odometer Attester API
  version: "1.0"
paths:
...
  /.well-known/nsm-attestation:
    get:
      consumes:
      - application/json
      description: Get the Nitro Security Module attestation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nitrite.Result'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app.codeResp'
      summary: Get NSM attestation
      tags:
      - attestation
...
  /vehicle/{tokenId}:
    get:
      consumes:
      - application/json
      description: Get information about a specific vehicle by token ID
      parameters:
      - description: Vehicle Token ID
        in: path
        name: tokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_DIMO-Network_odometer-attester_internal_client_identity.GraphQLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_app.codeResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app.codeResp'
      summary: Get vehicle information
      tags:
      - vehicle
  /vehicle/{tokenId}/odometer:
    get:
      consumes:
      - application/json
      description: Get the odometer reading for a specific vehicle by token ID
      parameters:
      - description: Vehicle Token ID
        in: path
        name: tokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cloudevent.CloudEvent-encoding_json_RawMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_app.codeResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_app.codeResp'
      summary: Get vehicle odometer reading
      tags:
      - vehicle
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"